Applying manual fixup patch for rbm
fbshipit-source-id: 94d10ec80a19c7b883bc14ad8a7b66f11d761f9b

+++ arch/x86/lib/misc.c
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: GPL-2.0
 /*
  * Count the digits of @val including a possible sign.
  *
@@ -5,7 +6,7 @@
  */
 int num_digits(int val)
 {
-	int m = 10;
+	long long m = 10;
 	int d = 1;
 
 	if (val < 0) {
